@@ -14,7 +14,7 @@
   ],
   "require": {
     "pear/archive_tar": "^1.4.14",
-    "php": ">=7.4",
+    "php": ">=8.0",
     "ext-curl": "*",
     "ext-zip": "*",
     "ext-mbstring": "*",
